requirements.txt
appdirs==1.4.0
Django==1.9.5
django-crispy-forms==1.6.0
django-filter==1.0.1
django-markdown-deux==1.0.5
django-pagedown==0.1.1
djangorestframework==3.5.4
djangorestframework-jwt==1.9.0
Markdown==2.6.8
markdown2==2.3.1
packaging==16.8
Pillow==3.1.0
PyJWT==1.4.2
pyparsing==2.1.10
six==1.10.0
